:fire: Refs #306 - Remove unneeded code
