+++ src/openforms/submissions/tests/test_models.py
@@ -110,74 +110,6 @@ class TestSubmission(TestCase):
             "test1, test2",
         )
 
-    def test_submission_has_sensitive_information_returns_true_with_sensitive_information_in_form_definition(
-        self,
-    ):
-        form_definition = FormDefinitionFactory.create(
-            configuration={
-                "components": [
-                    {"key": "textFieldSensitive", "isSensitiveData": True},
-                    {"key": "textFieldNotSensitive", "isSensitiveData": False},
-                ],
-            }
-        )
-        form = FormFactory.create()
-        form_step = FormStepFactory.create(form=form, form_definition=form_definition)
-
-        submission = SubmissionFactory.create(form=form)
-        SubmissionStepFactory.create(
-            submission=submission,
-            data={
-                "textFieldSensitive": "this is sensitive",
-                "textFieldNotSensitive": "this is not sensitive",
-            },
-            form_step=form_step,
-        )
-
-        self.assertTrue(submission.has_sensitive_information)
-
-    def test_submission_has_sensitive_information_returns_false_when_no_sensitive_information_in_form_definition(
-        self,
-    ):
-        form_definition = FormDefinitionFactory.create(
-            configuration={
-                "components": [
-                    {"key": "textFieldNotSensitive", "isSensitiveData": False},
-                ],
-            }
-        )
-        form = FormFactory.create()
-        form_step = FormStepFactory.create(form=form, form_definition=form_definition)
-
-        submission = SubmissionFactory.create(form=form)
-        SubmissionStepFactory.create(
-            submission=submission,
-            data={
-                "textFieldNotSensitive": "this is not sensitive",
-            },
-            form_step=form_step,
-        )
-
-        self.assertFalse(submission.has_sensitive_information)
-
-    def test_submission_has_sensitive_information_returns_true_if_bsn_in_submission(
-        self,
-    ):
-        submission = SubmissionFactory.create(bsn="999990676")
-        self.assertTrue(submission.has_sensitive_information)
-
-    def test_submission_has_sensitive_information_returns_true_if_kvk_in_submission(
-        self,
-    ):
-        submission = SubmissionFactory.create(kvk="69599084")
-        self.assertTrue(submission.has_sensitive_information)
-
-    def test_submission_has_sensitive_information_returns_false_if_already_cleaned(
-        self,
-    ):
-        submission = SubmissionFactory.create(_is_cleaned=True)
-        self.assertFalse(submission.has_sensitive_information)
-
     def test_submission_remove_sensitive_data(self):
         form_definition = FormDefinitionFactory.create(
             configuration={
